+++ test/union.js
@@ -8,26 +8,26 @@ var Point = t.struct({
   y: t.Number
 });
 
+var Circle = t.struct({
+  center: Point,
+  radius: t.Number
+}, 'Circle');
+
+var Rectangle = t.struct({
+  a: Point,
+  b: Point
+});
+
+var Shape = t.union([Circle, Rectangle], 'Shape');
+
+Shape.dispatch = function (values) {
+  assert(t.Object.is(values));
+  return values.hasOwnProperty('center') ?
+    Circle :
+    Rectangle;
+};
+
 describe('t.union(types, [name])', function () {
-
-  var Circle = t.struct({
-    center: Point,
-    radius: t.Number
-  }, 'Circle');
-
-  var Rectangle = t.struct({
-    a: Point,
-    b: Point
-  });
-
-  var Shape = t.union([Circle, Rectangle], 'Shape');
-
-  Shape.dispatch = function (values) {
-    assert(t.Object.is(values));
-    return values.hasOwnProperty('center') ?
-      Circle :
-      Rectangle;
-  };
 
   describe('combinator', function () {
 
@@ -54,25 +54,6 @@ describe('t.union(types, [name])', function () {
   });
 
   describe('constructor', function () {
-
-    it('should have a default dispatch() implementation', function () {
-      var T = t.union([t.String, t.Number], 'T');
-      assert.deepEqual(T(1), 1);
-    });
-
-    it('should have a default dispatch() implementation handling union of unions', function () {
-      var T1 = t.union([t.String, t.Number], 'T1');
-      var T2 = t.union([t.Boolean, t.Object], 'T2');
-      var T = t.union([T1, T2], 'T');
-      assert.deepEqual(T({foo: "bar"}), {foo: "bar"});
-    });
-
-    it('should throw when dispatch() does not return a type', function () {
-      throwsWithMessage(function () {
-        var T = t.union([t.String, t.Number], 'T');
-        T(true);
-      }, '[tcomb] Invalid value true supplied to T');
-    });
 
     it('should build instances when dispatch() is implemented', function () {
       var circle = Shape({center: {x: 0, y: 0}, radius: 10});
@@ -108,6 +89,40 @@ describe('t.union(types, [name])', function () {
     it('should return true when x is an instance of the union', function () {
       var p = new Circle({center: { x: 0, y: 0 }, radius: 10});
       assert.ok(Shape.is(p));
+    });
+
+  });
+
+  describe('#dispatch(x)', function () {
+
+    it('should have a default implementation', function () {
+      var T = t.union([t.String, t.Number], 'T');
+      assert.deepEqual(T(1), 1);
+    });
+
+    it('should handle union of unions', function () {
+      var T1 = t.union([t.String, t.Number], 'T1');
+      var T2 = t.union([t.Boolean, t.Object], 'T2');
+      var T = t.union([T1, T2, t.Array], 'T');
+      assert.strictEqual(T.dispatch(1), t.Number);
+      assert.strictEqual(T.dispatch({foo: "bar"}), t.Object);
+      assert.strictEqual(T.dispatch([]), t.Array);
+    });
+
+    it('should throw if does not return a type', function () {
+      throwsWithMessage(function () {
+        var T = t.union([t.String, t.Number], 'T');
+        T(true);
+      }, '[tcomb] Invalid value true supplied to T');
+    });
+
+  });
+
+  describe('#update(instance, spec)', function () {
+
+    it('should update the right instance', function () {
+      var circle = Shape.update({ center: { x: 0, y: 0 }, radius: 10 }, { radius: { $set: 15 } });
+      assert.strictEqual(Circle.is(circle), true);
     });
 
   });